+++ lib/src/widgets/frames.dart
@@ -66,7 +66,16 @@ class FramesWidgetState extends State<Frames> {
     _apiService = CheckoutApiService(
       publicKey: widget.config.publicKey,
       debug: widget.config.debug,
+      enableLogging: widget.config.enableLogging,
     );
+    
+    // Log frames initialization
+    _apiService.logger.logFramesInit();
+    
+    // Log frames ready after next frame
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _apiService.logger.logFramesReady();
+    });
   }
 
   @override